refactor: remove re-export indirection, point imports at source modules
Each function was imported and then re-exported through a barrel, creating
duplicate import/export for the same name in the same file. Fix by removing
the barrel re-exports and updating callers to import directly from the
module that defines the function.
